Hashed out some key formatting for the reagent results table and included the colorpicker from Bootstrap to select colors for each of the reagent results.
As of this version, the main form (addquestion.html right now, will change in the near future to what it's supposed to be, add_data.html) not only populates the associated fields as individual <select> tags by pulling relevant data from the proper tables in the database, but, it now also creates a row of 3 colorpickers for each entry located in the Reagent table. the ids for those inputs are (Reagent.name)id1, (Reagent.name)id2, (Reagent.name)id3.  The picker itself is (Reagent.name)picker1, etc etc. For example, Marquisid1, Marquispicker1....

I feel accomplished!