Updated README...
Currently trying to draw up new tables for changes due to database schema
Previously colors were stored as strings and thought about changing them
to hexcodes and/or use a color picker of some sort.

What I'm aiming to change to a drop down that only has the colors that
reaction normally encounters. For example, Ehrlich, turns purple in the
presense of an indole. Ehrlich 99 times out of 100 does one of two things:

It turns purple. Or it doesn't. That's it. Why mess with a color picker,
and pick out a shade of purple that due to people eyes, cellphone color
variations and mess of other factors probably isn't even remotely accurate
anyway. It'll take time to do so as well.  If it's off? Always add 'Other'
at the last spot and recorders can just add in what happened. (probably
tack that into the questions table).

This allows me a built in way to search the database for anything that
comes up 'other' for further information. Being able to look for outliers
easily is essential.

Not all of the tables are modified to use this new schema ( nor have I
tested it yet at all ).