More radically different changes to the ovreall database schema
With this, I should be satisfied for a bit and see no reason to
make any additional changes.

Please note, that this commit -only- changes SQLAlchemy's declarative
classes that both create the tables in the database itself, and,
provide those tables as class objects to be used in the rest of the
app.

Everything else, as of right this second, should be broken. Will
start on reformulating individual pages and probably massively
overhauling them too, in the next 12 hours or so.