Wrote up /AddReagentList to the list of app.routes.  Definitely struggled for a while on this.
Since I'm still not really 100% comfortable in my ability to create/identify/work-inside/manipulate data structures ( I tend to just stick with dicts, though I'm learning that this isn't always the best idea...) I'm trying to expand my horizons a bit.  So I set a bit of a handicap if you will, and just put all of the reagents into a list and imported them over, then, made it a list of lists, and stuffed it into the DatabaseMasterList.py.

This file is likely what I'm going to keep various collections of default data that way I can easily readd when I change something in the database schema.

I thought I was going to have one more shot approute without having to worry about the UUIDs directly, but I totally forgot that the reagents table has an author, and that author column, is a UUID.

I started off by importing the list over, and iterating over it, pulling the full name from that dataset, since it's all me for right now, I just went with the first one.

What this Query did though, is return the associated UUID with that full name.  This is something I had done before. Good times.

It was even -more- fun when I finally learned that I hadn't, in fact, put my own name in the Users table. THings went a -slightly- more smoothly after that. But not by much.

From there it was just iterating over the list-
TODO: Get that god forsaken authorUUID Query OUT OF THE LOOP. That only needs to run once.
iterate over the list, inserting the needed data into the Reagents table by way of line 129, then adding that to the session, then going to the next reagent ( and making that worthless call to figure out the UUID again...) Once it iterates over everything, commit the session.  Cool.

Originally, the plan was to work out what tables I needed to populate to make this structure work, and I wasn't going to have it output anything at all, but apparently Flask doesn't really like that. It -really- prefers that it sends -something- nback to the user. Makes sense.

So I figured I could solve 2 birds with one stone, see if I can get any better at these html Bootstrap pages I'm going to use for all the menus, while I practice on being able to exact this data. I struggled for quite sometime with bootstrap initially having -zero- webdesign interest or knowledge, other than HTML from like the mid ninties. Learned a lot. Feel a bit better about it. Check the templates folder if you want.

The later half took even longer. Because see, the Reagents.author now stores a UUID, and that's not exactly friendly looking now is it?

But I want to know who that belongs to. ( at least in this particular instance ). So I decided to sit down and figure out how to make it present the username back again, 'cause I know as I move forward with this project, more and more of this will be dealing will be comparing and manipulating UUIDs all over the freakin' place.

In any case, mission accomplished!