Finished up the addcolorreactions URL @app.route's front end, and back end.
This special URL may be utilized in the future when we decide to add more substances to the list and get color reactions for them. Have one page left to go the main survey page, I'll sitting down and writing out a lot of documentation for this when I have more time over the next couple of days.