/static/img/reagents/%reagentname%/ created for storage of newly created, and descriptively named, images for reference.
Now, instead of using the color system I was attempting to write up, ( and had crazy long unfiltered color lists inside of ), we have what we've all come to know and love. The actual colors from the chart.

Images were taken from a high rez ( 3000 x 2400 ) digital copy. Got use GIMP for the first time, decided to go with that over anything else I have access to because of it's Python-Fu scripting support.  Made quick work of that file, but didn't let me name the images what I wanted to, so instead I got around 700~ slice_XX_XX.png files. With 66% of them being the board I cut out from the rest.

Instead of just throwing em onto the server and linking to them ( there were sequential, so it's relatively easy to know what goes where... ) It would have left them with no descriptions.  Since one of the largest concerns for me in this overall project it's long term viability and use, part of that, is the possibility of hosting reagent results online.  Even if it's just my personal server for now.

In switching from the old way of doing color to this greatly enhanced version, I also came to realize an incredibly strong advantage to doing so: I've effectively made it incredibly easy for us to create color reagent reactions from things that didn't quite make it onto our existing chart.

It's not that we didn't want more info, there's only so much space.  And printing physical in at these volumes, especially because of the attention to paid to making sure we have the most accurate representation of that color as possible. It's expensive. Because of this, we only have 29 substances on the standard table, along with another 12 substances on the back for that may or may not react with Ehrlich.

This opens up a new avenue for us, and allows us the ability to store -many- more times the reaction combinations and in turn, should allow us to test for the presence of most substances we've ever been able to test for previously.  However this also translates to into there being no physical media for these reactions, especially if we get analysis from EcstacyData and others.  People may want to save those reactions to their devices for later reference.  Now instead of trying to figure out what reaction slice_46_8.png belongs to, well... what's it a name anyway right?

Cheers.

<3